FIX: Inconsistent table rename broke the plugin
It looks like `akismet_data_fields` was renamed to `akismet_spam_data`
but no migration was present for this. Additionally, the filename was
still `akisment_data_fields.rb` which caused the plugin the break.

This commit adds a migration to rename the table and renames the ruby
file so the model can be found.