FIX: We don't want to double-process flagged posts. (#62)
There's a chance a TL3 user flags a TL0 post as spam before Akismet has a chance of processing it. If it happens, only a flagged post approved by the system will appear in the queue. Staff won't be able to restore it from the queue because Akismet already deleted it (should be hidden instead).