FIX: Ensure same entities are not checked twice
Multiple instances of the same scheduled job can run at the same time
if the execution time of an instance is longer than the schedule
interval.

The list of checked entities is fetched when the job begins which
combined with multiple instances resulted in the same entities being
checked multiple times.