FIX: Ensure post is checked when raw changes
This solves two edge cases. The first one is the basic one when the
user simply edits the post, which did not trigger a re-check of new
raw.

The other case is more complicated and exploits the fact that deleted
posts are skipped. Users can post spam, quickly delete the post, wait
for the post to be processed and skipped because it is deleted and then
restore it with the spam content.