FEATURE: Various improvements and refactoring (#32)
* FEATURE: Show Akismet status in flag post modal

This can help understand why some posts were marked as spam or not by
Akismet.

* DEV: Move all styles into a single stylesheet

Both previous stylesheets combined have less than 15 lines together.

* FEATURE: Send feedback for posts belonging to spammers

* DEV: Prefer require_relative to load dependencies

* FIX: Ensure post is checked when raw changes

This solves two edge cases. The first one is the basic one when the
user simply edits the post, which did not trigger a re-check of new
raw.

The other case is more complicated and exploits the fact that deleted
posts are skipped. Users can post spam, quickly delete the post, wait
for the post to be processed and skipped because it is deleted and then
restore it with the spam content.

* FIX: Resolve all reviewables of a user

Acting on an Akismet reviewable did not automatically update the status
of all other reviewable or submit feedback.

* DEV: Rename state new to pending

* UX: Add descriptions to reviewable actions

* DEV: Remove Excon stubs

Excon stubs used to leak between tests.