FIX: Make the review of a TL1 user's first post configurable.
This check is important because TL1 is the only trust level that doesn't require visiting the forum multiple days. Combine this with the plugin's default settings, and a spammer could bypass Akismet by reading a couple of posts
before creating their own.

However, as pointed out in #51, the check is invisible to staff members and can't be disabled, even by setting `skip_akismet_trust_level` to 0.